:sparkles: :: Add major toText
