+++ core/model/src/main/java/com/goms/model/enum/Major.kt
@@ -5,11 +5,3 @@ enum class Major(val value: String) {
     SMART_IOT("IoT"),
     AI("AI")
 }
-
-fun Major.toText(): String {
-    return when (this) {
-        Major.SW_DEVELOP -> "SW개발"
-        Major.SMART_IOT -> "IoT"
-        Major.AI -> "AI"
-    }
-}
