+++ feature/main/src/main/java/com/goms/main/component/MainLateCard.kt
@@ -35,8 +35,8 @@ import com.goms.design_system.component.modifier.gomsClickable
 import com.goms.design_system.theme.GomsTheme
 import com.goms.main.viewmodel.GetLateRankListUiState
 import com.goms.model.enum.Authority
-import com.goms.model.enum.toText
 import com.goms.model.response.late.RankResponse
+import com.goms.ui.toText
 
 @Composable
 fun MainLateCard(
